requirements.txt
datasets
humanfriendly
kapral>=0.1.2
langdetect
lightning
protonn
pytorch_lightning
transformers
vecto
wandb
